+++ common/protocol/src/methods/checks/isPoolActive.ts
@@ -17,7 +17,6 @@ export function isPoolActive(this: Validator): boolean {
     case PoolStatus.POOL_STATUS_ACTIVE:
       return true;
     case PoolStatus.POOL_STATUS_NO_FUNDS:
-      this.logger.warn("Pool is out of funds, rewards may be reduced");
       return true;
     case PoolStatus.POOL_STATUS_DISABLED:
       this.logger.info(
@@ -36,6 +35,9 @@ export function isPoolActive(this: Validator): boolean {
       return false;
     case PoolStatus.POOL_STATUS_UNSPECIFIED:
       this.logger.info("Pool status is currently unspecified. Idling ...");
+      return false;
+    case PoolStatus.POOL_STATUS_END_KEY_REACHED:
+      this.logger.info("End key reached. Idling ...");
       return false;
     default:
       this.logger.info("Pool status is currently unknown. Idling ...");