+++ common/sdk/src/amino/delegation.amino.ts
@@ -11,7 +11,7 @@ import { isNotEmpty } from "../utils";
 export const createDelegationAminoConverters = (): AminoConverters => {
   return {
     "/kyve.delegation.v1beta1.MsgDelegate": {
-      aminoType: "kyve/delegation/MsgDelegate",
+      aminoType: "/kyve.delegation.v1beta1.MsgDelegate",
       toAmino: (msg: MsgDelegate) => ({
         creator: msg.creator,
         staker: msg.staker,
@@ -24,7 +24,7 @@ export const createDelegationAminoConverters = (): AminoConverters => {
       }),
     },
     "/kyve.delegation.v1beta1.MsgWithdrawRewards": {
-      aminoType: "kyve/delegation/MsgWithdrawRewards",
+      aminoType: "/kyve.delegation.v1beta1.MsgWithdrawRewards",
       toAmino: (msg: MsgWithdrawRewards) => ({
         creator: msg.creator,
         staker: msg.staker,
@@ -35,7 +35,7 @@ export const createDelegationAminoConverters = (): AminoConverters => {
       }),
     },
     "/kyve.delegation.v1beta1.MsgRedelegate": {
-      aminoType: "kyve/delegation/MsgRedelegate",
+      aminoType: "/kyve.delegation.v1beta1.MsgRedelegate",
       toAmino: (msg: MsgRedelegate) => ({
         creator: msg.creator,
         from_staker: msg.from_staker,
@@ -50,7 +50,7 @@ export const createDelegationAminoConverters = (): AminoConverters => {
       }),
     },
     "/kyve.delegation.v1beta1.MsgUndelegate": {
-      aminoType: "kyve/delegation/MsgUndelegate",
+      aminoType: "/kyve.delegation.v1beta1.MsgUndelegate",
       toAmino: (msg: MsgUndelegate) => ({
         creator: msg.creator,
         staker: msg.staker,